refactor: config 책임 분리
